js and css setup
